fix: use injected acorn instance
Since rollup bundles it's own version of acorn `require("acorn")` will never be the same as `Parser.acorn` and you always gets the error "acorn-private-class-elements does not support mixing different acorn copies". If this change looks good, I can make a similar one for acorn-class-fields (and also bump it to use latest version of this module).