Migrate to ESM
We've recently migrated Chrome DevTools to ESM, and have traditionally
been using acorn as the parser for various features including the
prettifier. In the context of https://crbug.com/1084349 we are looking
into including private fields support in Chrome DevTools via this
plugin, and in order to make that easier with our build system, I've
added support for ESM output here in addition to the CJS output.

Similar change to
https://github.com/acornjs/acorn-private-methods/pull/4.